Move `configure_logging` to its own file
